detection/requirements.txt
# requirements
# Usage: pip install -r requirements.txt

numpy
pandas
pika
ultralytics
opencv-python
PyYAML
pillow